fix casting of image quality

+++ Magic of East/app/Trait/UploadImage.php
@@ -15,7 +15,7 @@ trait UploadImage
                 $image = Image::read($image)->resize(null, null, function ($constraint) {
                     $constraint->aspectRatio();
                     $constraint->upsize();
-                })->encode(new AutoEncoder(quality: ((100 * 2100000) / $image->getSize())))
+                })->encode(new AutoEncoder(quality: (int)((100 * 2100000) / $image->getSize())))
                     ->save(public_path('storage/images/' . $image_name));
             } else {
                 $image->storeAs('images', $image_name, 'public');
